put in statement to only delete LCtemp if it exists

+++ preprocess/processData.py
@@ -145,7 +145,8 @@ class Landsat(object):
         '-te', '%f' % self.ulx, '%f' % self.lry,'%f' % self.lrx,'%f' % self.uly,\
         '-ts', '%f' % self.ncol, '%f' % self.nrow,'-multi','-of','GTiff','%s' % outfile, '%s' % outfile2]
         warp(optionList)
-        shutil.rmtree(LCtemp)
+        if os.path.exists(LCtemp):
+            shutil.rmtree(LCtemp)
